Add a nixos container that can run full cluster
This PR when merged will allow you to run cluster setup almost exactly
like it does in production, in a matter of seconds. The word
"container" here may be a bit misleading, because even though we do
start a cgroup, it will run systemd, nild, nginx, and other things
like a normal operating system (and unlike Docker for example).

Note, that you need NixOS installed and not just Nix package manager,
because the `nixos-container` tool is not available with just Nix.

```
sudo nixos-container create devcontainer --flake .
sudo nixos-container start devcontainer
```

If you made changes to nild or configs and want to update the
container in-place, you can do:

```
sudo nixos-container update devcontainer --flake .
```

As a follow-up, I will add ClickHouse, exporter, and all frontend
parts to run in the container as well.